pyproject.toml
[project]
name = "schoolflix"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aerich>=0.9.1",
    "aioredis>=2.0.1",
    "argon2-cffi>=25.1.0",
    "fastapi[all]>=0.118.0",
    "python-jose[cryptography]>=3.5.0",
    "redis>=6.4.0",
    "tortoise-orm>=0.25.1",
]

[tool.aerich]
tortoise_orm = "app.config.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.poetry.scripts]
migrate = "aerich migrate"
upgrade = "aerich upgrade"
